Refactor trait to allow outputting other types
