+++ src/lib.rs
@@ -1,14 +1,7 @@
 #![doc = include_str!("../README.md")]
 #![deny(clippy::invalid_regex)]
 
-use std::{
-    borrow::Cow,
-    collections::HashSet,
-    marker::PhantomData,
-    ops::{Range, RangeInclusive},
-    path::Path,
-    str::Utf8Error,
-};
+use std::{collections::HashSet, marker::PhantomData, path::Path, str::Utf8Error};
 
 use content::Content;
 use derivative::Derivative;
@@ -18,6 +11,7 @@ use flagset::{flags, FlagSet};
 use getset::{CopyGetters, Getters};
 use itertools::Itertools;
 use once_cell::sync::OnceCell;
+use parser::bytes::Location;
 use strum::{Display, EnumIter};
 use tap::{Conv, Pipe};
 use thiserror::Error;
@@ -40,12 +34,13 @@ pub mod text;
 /// [`Extractor`]: crate::Extractor
 /// [`Error`]: enum@crate::Error
 pub mod impl_prelude {
+    pub use super::parser::bytes::Location as SnippetLocation;
     pub use super::{
         Context as SnippetContext, Error, Extractor as SnippetExtractor, Kind as SnippetKind,
         Kinds as SnippetKinds, Language as SnippetLanguage, LanguageError,
-        Location as SnippetLocation, Metadata as SnippetMetadata, Method as SnippetMethod,
-        Options as SnippetOptions, Snippet, Strategy as LanguageStrategy, Target as SnippetTarget,
-        Transform as SnippetTransform, Transforms as SnippetTransforms,
+        Metadata as SnippetMetadata, Method as SnippetMethod, Options as SnippetOptions, Snippet,
+        Strategy as LanguageStrategy, Target as SnippetTarget, Transform as SnippetTransform,
+        Transforms as SnippetTransforms,
     };
 }
 
@@ -76,20 +71,20 @@ impl Error {
 #[error("language: {0}")]
 pub struct LanguageError(String);
 
-/// An implementation of [`Extractor`] enables snippets to be extracted
+/// An implementation of [`Extractor`] enables output to be extracted
 /// from a given unit of source code (typically a file).
 pub trait Extractor {
     /// The options accepted by the extractor.
     type Options;
 
-    /// The snippet type returned by the extractor.
-    type Snippet;
+    /// The output of the extractor.
+    type Output;
 
-    /// Reads the provided unit of source code for snippets, according to the provided options.
-    fn extract(opts: &Self::Options, content: &Content) -> Result<Vec<Self::Snippet>, Error>;
+    /// Reads the provided unit of source code for output, according to the provided options.
+    fn extract(opts: &Self::Options, content: &Content) -> Result<Self::Output, Error>;
 
-    /// Reads the provided file for snippets, according to the provided options.
-    fn extract_file(opts: &Self::Options, path: &Path) -> Result<Vec<Self::Snippet>, Error> {
+    /// Reads the provided file for output, according to the provided options.
+    fn extract_file(opts: &Self::Options, path: &Path) -> Result<Self::Output, Error> {
         let content = Content::from_file(path).map_err(Error::ReadFile)?;
         Self::extract(opts, &content)
     }
@@ -456,177 +451,6 @@ pub struct Metadata {
 impl std::fmt::Display for Metadata {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         write!(f, "{}/{}/{}", self.kind, self.method, self.location)
-    }
-}
-
-/// The location in the unit of source code from which the snippet was extracted.
-///
-/// After opening the file (so a hypothetical reader is at byte offset `0`),
-/// the reader then skips a number of bytes equal to `byte_offset`,
-/// then reads a number of bytes equal to `byte_len`.
-/// The bytes that were read compose the entire snippet.
-///
-/// For example, given the file:
-/// ```not_rust
-/// #include <stdio.h>
-///
-/// int main() {
-///   printf("hello world\n");
-///   return 0;
-/// }
-/// ```
-///
-/// In the representation the computer sees, it looks like this (using `⏎` to represent a newline):
-/// ```not_rust
-/// #include <stdio.h>⏎⏎int main() {⏎  printf("hello world\n");⏎  return 0;⏎}⏎
-/// ^^^^                ^        ^
-/// 0123...             20 <-9-> 29
-/// ```
-///
-/// The [`Location`] below represents the `int main()` snippet in that example:
-/// ```
-/// # // ⏎ is a multi-byte symbol, so use an empty space for demonstration instead.
-/// # let example = "#include <stdio.h>  int main() {}";
-/// # use snippets::*;
-/// let location = Location::builder().byte_offset(20).byte_len(10).build();
-///
-/// let range = location.as_range();
-/// let snippet = &example.as_bytes()[range];
-///
-/// let got = std::str::from_utf8(snippet)?;
-/// assert_eq!(got, "int main()");
-/// # Ok::<(), std::str::Utf8Error>(())
-/// ```
-//
-// Note: we use a `TypedBuilder` instead of a `Constructor` here because this way we can accept
-// a standard `usize` for each argument while still making it very clear in-code
-// which argument is which.
-//
-// Basically, the intent is to straddle the line between newtype convenience and newtype safety.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, PartialOrd, Ord, CopyGetters, TypedBuilder)]
-#[getset(get_copy = "pub")]
-pub struct Location {
-    /// The byte offset at which the snippet began.
-    #[builder(setter(transform = |input: usize| ByteOffset(input)))]
-    byte_offset: ByteOffset,
-
-    /// The number of bytes to read for the snippet from the file.
-    #[builder(setter(transform = |input: usize| ByteLen(input)))]
-    byte_len: ByteLen,
-}
-
-impl std::fmt::Display for Location {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}..={}", self.start_byte(), self.end_byte())
-    }
-}
-
-impl Location {
-    /// Read a [`Location`] as a range, intended to be used to index a buffer of bytes.
-    pub fn as_range(&self) -> std::ops::Range<usize> {
-        let start = self.byte_offset.0;
-        let len = self.byte_len.0;
-        let end = start + len;
-        start..end
-    }
-
-    /// The index of the first byte indicated for the provided location.
-    pub fn start_byte(&self) -> usize {
-        self.as_range().start
-    }
-
-    /// The index of the last byte indicated for the provided location.
-    pub fn end_byte(&self) -> usize {
-        let end = self.as_range().end;
-        if end == 0 {
-            0
-        } else {
-            end - 1 // as_range is not inclusive, so the last byte _to be read_ is less one.
-        }
-    }
-
-    /// Extract the bytes indicated by a [`Location`] from a buffer.
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use snippets::*;
-    /// let example = "#include <stdio.h>  int main() {}";
-    /// let location = Location::builder().byte_offset(20).byte_len(10).build();
-    ///
-    /// let got = location.extract_from(example.as_bytes());
-    /// assert_eq!(got, b"int main()");
-    /// ```
-    pub fn extract_from<'a>(&self, buf: &'a [u8]) -> &'a [u8] {
-        &buf[self.as_range()]
-    }
-
-    /// Extract the bytes indicated by a [`Location`] from a buffer,
-    /// into a lossily converted [`String`].
-    ///
-    /// # Example
-    ///
-    /// ```
-    /// # use snippets::*;
-    /// let example = "#include <stdio.h>  int main() {}";
-    /// let location = Location::builder().byte_offset(20).byte_len(10).build();
-    ///
-    /// let got = location.extract_from_lossy(example.as_bytes());
-    /// assert_eq!(got, "int main()");
-    /// ```
-    pub fn extract_from_lossy<'a>(&self, buf: &'a [u8]) -> Cow<'a, str> {
-        let bytes = self.extract_from(buf);
-        String::from_utf8_lossy(bytes)
-    }
-}
-
-impl From<Range<usize>> for Location {
-    fn from(value: Range<usize>) -> Self {
-        let start = value.start;
-        let end = value.end;
-        Self {
-            byte_offset: ByteOffset(start),
-            byte_len: ByteLen(end - start),
-        }
-    }
-}
-
-impl From<RangeInclusive<usize>> for Location {
-    fn from(value: RangeInclusive<usize>) -> Self {
-        let start = *value.start();
-        let end = *value.end() + 1;
-        Self {
-            byte_offset: ByteOffset(start),
-            byte_len: ByteLen(end - start),
-        }
-    }
-}
-
-/// The byte offset at which the snippet began.
-///
-/// Zero-based, meaning that if the snippet begins on the first byte of the file,
-/// this offset is `0`.
-///
-/// Think of the offset as
-/// "the number of bytes to skip from the start of the file to when this snippet begins".
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, PartialOrd, Ord, derive_more::Display)]
-pub struct ByteOffset(usize);
-
-impl ByteOffset {
-    /// View the offset as a usize.
-    pub fn as_usize(self) -> usize {
-        self.0
-    }
-}
-
-/// The number of bytes to read for the snippet from the file.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, PartialOrd, Ord, derive_more::Display)]
-pub struct ByteLen(usize);
-
-impl ByteLen {
-    /// View the length as a usize.
-    pub fn as_usize(self) -> usize {
-        self.0
     }
 }
 
@@ -1430,10 +1254,7 @@ mod tests {
 
     #[test]
     fn location_round_trip() {
-        let location = Location {
-            byte_offset: ByteOffset(20),
-            byte_len: ByteLen(10),
-        };
+        let location = Location::builder().byte_offset(20).byte_len(10).build();
         let inclusive_range = 20..=29;
         let range = 20..30;
 
@@ -1444,10 +1265,7 @@ mod tests {
 
     #[test]
     fn location_bytes_round_trip() {
-        let location = Location {
-            byte_offset: ByteOffset(10),
-            byte_len: ByteLen(10),
-        };
+        let location = Location::builder().byte_offset(10).byte_len(10).build();
 
         let input = "0123456789helloworld_abcdefghijk";
         assert_eq!(location.extract_from(input.as_bytes()), b"helloworld");
@@ -1466,10 +1284,7 @@ mod tests {
     #[test]
     fn location_extract() {
         let input = "0123456789helloworld0123456789";
-        let location = Location {
-            byte_offset: ByteOffset(10),
-            byte_len: ByteLen(10),
-        };
+        let location = Location::builder().byte_offset(10).byte_len(10).build();
 
         assert_eq!(location.extract_from(input.as_bytes()), b"helloworld");
         assert_eq!(location.start_byte(), 10);
@@ -1479,10 +1294,7 @@ mod tests {
     #[test]
     fn location_extract_end() {
         let input = "0123456789helloworld";
-        let location = Location {
-            byte_offset: ByteOffset(10),
-            byte_len: ByteLen(10),
-        };
+        let location = Location::builder().byte_offset(10).byte_len(10).build();
 
         assert_eq!(location.extract_from(input.as_bytes()), b"helloworld");
     }
